pyproject.toml
[project]
name = "midom"
version = "0.1.0"
description = "Concepts and relations to describe the process of DICOM deidentification"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.11.5",
    "pydicom>=2.4.4",
    "sphinx>=7.4.7",
    "sphinx-rtd-theme>=3.0.2",
    "sphinxcontrib-plantuml>=0.30",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "livereload>=2.7.1",
    "pytest>=8.3.5",
    "setuptools>=78.1.0",
]

[tool.black]
line-length = 79